move ac to action/ from skill/
